updated the file structure and made it more maintainable (naming could use work) and able to grow as we add more.
