+++ src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LandingPage from "./components/LandingPage";
+import LandingPage from "./components/LandingPage/LandingPage";
 
 function App() {
   return (
