+++ src/components/LandingPage/LandingPage.css
@@ -1,0 +1,38 @@
+.landing-page {
+    position: relative;
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    background: linear-gradient(to bottom, #4fd1c5, #3b82f6);
+    overflow: hidden;
+    font-family: "Cormorant Garamond", serif;
+  }
+  
+  .content {
+    position: relative;
+    z-index: 10;
+    text-align: center;
+  }
+  
+  .title {
+    font-size: 4rem;
+    font-weight: bold;
+    margin-bottom: 1rem;
+  }
+  
+  .subtitle {
+    font-size: 1.5rem;
+    margin-bottom: 2rem;
+    color: white;
+  }
+  
+  @media (max-width: 640px) {
+    .title {
+      font-size: 3rem;
+    }
+    
+    .subtitle {
+      font-size: 1.25rem;
+    }
+  }
